Switch PYANG to be installed for Python3
Closes VYINT-5111

Discovered when running the DRAM acceptancte tests that pyang was being installed as a python2 dist-package, which was causing errors.  This diffset converts the DANOS:pyang build to python3.  Once this is approved, and I've checked there are no issues with the ISO, I will convert Vyatta:Tools to distribute from DANOS not VC (but if anyone needs the python2 version, they can get it from VC).